driver ping test along with some infra piece (#5424)
* Changes for running driverPing.go in turin env

* fixing golint errors

* updating lif qstate address as it changed after pulling from master

* keeping the nlog level as info for model_server

* moved chmod out of inner if to get it executed when the user is root

* linter error fix

+++ nic/sim/naples/start-naples-docker.sh
@@ -18,7 +18,7 @@ docker rmi -f naples:$VER
 echo "Loading docker image into registry ..."
 docker load --input naples-docker-$VER.tgz
 
-if [ "$1" == "--qemu" ]
+if [ $# -eq 1 ] && [ "$1" == "--qemu" ]
   then
     echo "Running the NAPLES container in Qemu mode ..."
     docker run --rm -d --name naples-$VER --privileged -ti -p $AGENT_PORT:$AGENT_PORT  -p $HAL_PORT:$HAL_PORT --mount type=bind,source=$NAPLES_DIR,target=/naples/data -e WITH_QEMU=1 "$CONTAINER"
